fix(tests): Split step tests to own file   * Updated runner describe titles   * Added maxWorkers to test script
